refactor(opencode): extract shared CLI tool renderers
The new minimal-mode REPL (`repl.ts`) and the existing one-shot `run`
command had ~200 lines of identical helpers (block/inline/fallback
plus per-tool renderers). Move them into `cli/cmd/render.ts` and
expose three entry points: `inline`, `renderTool`, `renderRunningTask`.

Pure code-move, no behavior change.
